+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * [#112](https://github.com/onechiporenko/my-dungeons-book/issues/112) Show items level for each equiped item in the Roster/Equipment
 * [#110](https://github.com/onechiporenko/my-dungeons-book/issues/110) New CN localizations
 * Add report-menu for heal and damage done by each spell
+* [#113](https://github.com/onechiporenko/my-dungeons-book/issues/113) Add context menu for Challenges table
 
 ### v2.4.0